Write locking now uses the flock system call to get an exclusive write lock on the OS
It no longer relies on the OS process pid that is in the write lock which causes issues on hard restart where there is stale lock files and pids get reused.

Add testing for locking in the same OS process and across OS processes.

Removed the exist clause in bitcask_lockops because existence of a lock file no longer proves that the DB is locked.
Now {error,locked} is returned from the nif if the DB is locked.

When the attempt to obtain the file write lock has failed, close the file descriptor.

wday-contrib 174