Delete hintfiles when corrupt
* Hintfile will be deleted when found corrupt.
* `require_hint_crc` when set to true will now always validate on fold.
  And when set to false, will never validate.

Previous behaviour worked as follows:
`require_hint_crc` when true would check to see if a CRC was appended a hintfile.
This would only occur when a data file was completely expired.
It would not validate the hintfile and instead error out.
Because neither the data file or hintfile were deleted, this would continue until restart.

wday-contrib 1309